requirements.txt
asgiref
backports.zoneinfo
Django
django-cors-headers
djangorestframework
sqlparse
typing_extensions
tzdata
